+++ common/core/src/main/java/is/codion/common/property/PropertyValue.java
@@ -32,21 +32,6 @@ public interface PropertyValue<T> extends Value<T> {
 	String propertyName();
 
 	/**
-	 * Returns the underlying value, if the value is null then a {@link IllegalStateException} is thrown.
-	 * @return the value, if available
-	 * @throws IllegalStateException in case the underlying value is null
-	 */
-	T getOrThrow() throws IllegalStateException;
-
-	/**
-	 * Returns the underlying value, if the value is null then a {@link IllegalStateException} is thrown.
-	 * @param message the error message to use when throwing
-	 * @return the value, if available
-	 * @throws IllegalStateException in case the underlying value is null
-	 */
-	T getOrThrow(String message) throws IllegalStateException;
-
-	/**
 	 * Sets this value to null as well as removing it from the underlying store and clearing the system property.
 	 */
 	void remove();